Switch to PostgreSQL in all environments
allows us to use PostgreSQL-specific features (e.g. case-insensitive
tokenized searching)

+++ db/migrate/20170830184339_create_heroes.rb
@@ -1,7 +1,7 @@
 class CreateHeroes < ActiveRecord::Migration[5.1]
   def change
     create_table :heroes, :id => false do |t|
-      t.integer :id, null: false
+      t.bigint :id, null: false
       t.string :name, null: false
       t.integer :stars, null: false
       t.string :faction, null: false
