+++ db/migrate/20170831180632_create_materials.rb
@@ -3,7 +3,7 @@ class CreateMaterials < ActiveRecord::Migration[5.1]
     create_table :materials do |t|
       t.references :hero, index: true, null: false
       t.integer :count, null: false
-      t.integer :material_hero_id
+      t.bigint :material_hero_id
       t.integer :stars
       t.string :faction
     end
